On branch working2 	new file:   .vscode/tasks.json 	new file:   mainFile/Conference.cpp 	new file:   mainFile/Conference.hpp 	modified:   mainFile/ConferenceManagementFile.cpp 	modified:   mainFile/ConferenceManagementFile.exe 	new file:   mainFile/tempCodeRunnerFile.cpp
